# angular/tools/swagger/src/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gestion de stock REST API",
    "description": "Gestion de stock API documentation",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "JWT": []
    }
  ],
...
  "paths": {
...
    "/gestiondestock/v1/commandesclients/update/quantite/{idCommande}/{idLigneCommande}/{quantite}": {
      "patch": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "updateQuantiteCommande",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "quantite",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/etat/{idCommande}/{etatCommande}": {
      "patch": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "updateEtatCommande",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "etatCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "EN_PREPARATION",
                "VALIDEE",
                "LIVREE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/client/{idCommande}/{idClient}": {
      "patch": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "updateClient",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandesclients/update/article/{idCommande}/{idLigneCommande}/{idArticle}": {
      "patch": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "updateArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandeFournisseurs/update/quantite/{idCommande}/{idLigneCommande}/{quantite}": {
      "patch": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "updateQuantiteCommande_1",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "quantite",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandeFournisseurs/update/etat/{idCommande}/{etatCommande}": {
      "patch": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "updateEtatCommande_1",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "etatCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "EN_PREPARATION",
                "VALIDEE",
                "LIVREE"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandeFournisseurs/update/article/{idCommande}/{idLigneCommande}/{idArticle}": {
      "patch": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "updateArticle_1",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/ventes/{idVente}": {
      "get": {
        "tags": [
          "Ventes"
        ],
        "summary": "Rechercher une vente par ID",
        "description": "Cette méthode permet de rechercher une vente par son ID",
        "operationId": "findById",
        "parameters": [
          {
            "name": "idVente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La vente a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucune vente trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/ventes/{codeVente}": {
      "get": {
        "tags": [
          "Ventes"
        ],
        "summary": "Rechercher une vente par code",
        "description": "Cette méthode permet de rechercher une vente par son code",
        "operationId": "findByCode",
        "parameters": [
          {
            "name": "codeVente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "L'article a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucune vente trouvé avec le code fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VenteDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/utilisateurs/{idUtilisateur}": {
      "get": {
        "tags": [
          "Utilisateurs"
        ],
        "summary": "Rechercher un utilisateur par ID",
        "description": "Cette méthode permet de rechercher un utilisateur par son ID",
        "operationId": "findById_1",
        "parameters": [
          {
            "name": "idUtilisateur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "L'utilisateur a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucun utilisateur trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtilisateurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/fournisseurs/{idFournisseur}": {
      "get": {
        "tags": [
          "Fournisseurs"
        ],
        "summary": "Rechercher un fournisseur par ID",
        "description": "Cette méthode permet de rechercher un fournisseur par son ID",
        "operationId": "findById_2",
        "parameters": [
          {
            "name": "idFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Le fournisseur a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucun fournisseur trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/entreprises/{idEntreprise}": {
      "get": {
        "tags": [
          "Entreprises"
        ],
        "summary": "Rechercher une entreprise par ID",
        "description": "Cette méthode permet de rechercher une entreprise par son ID",
        "operationId": "findById_3",
        "parameters": [
          {
            "name": "idEntreprise",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "L'entreprise a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntrepriseDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucune entreprise trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntrepriseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesclients/lignesCommande/{idCommande}": {
      "get": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "findAllLignesCommandesClientByCommandeClientId",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LigneCommandeClientDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandeFournisseurs/{idCommandeFournisseur}": {
      "get": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "findById_4",
         "parameters": [
          {
            "name": "idCommandeFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandeFournisseurs/{codeCommandeFournisseur}": {
      "get": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "findByCode_1",
        "parameters": [
          {
            "name": "codeCommandeFournisseur",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandeClients/{idCommandeClient}": {
      "get": {
	"summary": "Rechercher une commande client par ID",
        "tags": [
          "CommandeClients"
        ],
        "operationId": "findById_5",
        "parameters": [
          {
            "name": "idCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/commandeClients/{codeCommandeClient}": {
      "get": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "findByCode_2",
        "parameters": [
          {
            "name": "codeCommandeClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/clients/{idClient}": {
      "get": {
        "tags": [
          "Clients"
        ],
        "summary": "Rechercher un client par ID",
        "description": "Cette méthode permet de rechercher un client par son ID",
        "operationId": "findById_6",
        "parameters": [
          {
            "name": "idClient",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Le client a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucun client trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/categories/id/{idCategory}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Rechercher une categorie par ID",
        "description": "Cette méthode permet de rechercher une categorie par son ID",
        "operationId": "findById_7",
        "parameters": [
          {
            "name": "idCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La categorie a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucune categorie trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
    "/gestiondestock/v1/categories/code/{codeCategory}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Rechercher une categorie par code",
        "description": "Cette méthode permet de rechercher une categorie par son code",
        "operationId": "findByCode_3",
        "parameters": [
          {
            "name": "codeCategory",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "La categorie a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucune categorie trouvé avec le code fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/articles/idArticle/{idArticle}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Rechercher un article par ID",
        "description": "Cette méthode permet de rechercher un article par son ID",
        "operationId": "findById_8",
        "parameters": [
          {
            "name": "idArticle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "L'article a été trouvé dans la BDD",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          },
          "404": {
            "description": "Aucun article trouvé dans la BDD avec l'ID fourni",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandesclients/delete/article/{idCommande}/{idLigneCommande}": {
      "delete": {
        "tags": [
          "CommandeClients"
        ],
        "operationId": "deleteArticle",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeClientDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/gestiondestock/v1/commandeFournisseurs/delete/article/{idCommande}/{idLigneCommande}": {
      "delete": {
        "tags": [
          "CommandeFournisseurs"
        ],
        "operationId": "deleteArticle_1",
        "parameters": [
          {
            "name": "idCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "idLigneCommande",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommandeFournisseurDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "VenteDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateVente": {
            "type": "string",
            "format": "date-time"
          },
          "commentaire": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AdresseDto": {
        "type": "object",
        "properties": {
          "adresse1": {
            "type": "string"
          },
          "adresse2": {
            "type": "string"
          },
          "ville": {
            "type": "string"
          },
          "codePostale": {
            "type": "string"
          },
          "pays": {
            "type": "string"
          }
        }
      },
      "EntrepriseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "codeFiscal": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "siteWeb": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/AdresseDto"
          }
        }
      },
...
      "UtilisateurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dateDeNaissance": {
            "type": "string",
            "format": "date-time"
          },
          "motDePasse": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "entreprise": {
            "$ref": "#/components/schemas/EntrepriseDto"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolesDto"
            }
          }
        }
      },
      "ArticleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "codeArticle": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "prixUnitaireHt": {
            "type": "number"
          },
          "tauxTva": {
            "type": "number"
          },
          "prixUnitaireTtc": {
            "type": "number"
          },
          "photo": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryDto"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "designation": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "FournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "adresse": {
            "$ref": "#/components/schemas/AdresseDto"
          }
        }
      },
      "CommandeFournisseurDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "etatCommande": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "fournisseur": {
            "$ref": "#/components/schemas/FournisseurDto"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
      "ClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nom": {
            "type": "string"
          },
          "prenom": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "numTel": {
            "type": "string"
          },
          "adresse": {
            "$ref": "#/components/schemas/AdresseDto"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "dateCommande": {
            "type": "string",
            "format": "date-time"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "etatCommande": {
            "type": "string",
            "enum": [
              "EN_PREPARATION",
              "VALIDEE",
              "LIVREE"
            ]
          },
          "client": {
            "$ref": "#/components/schemas/ClientDto"
          },
          "commandeLivree": {
            "type": "boolean"
          }
        }
      },
...
      "LigneCommandeClientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "quantite": {
            "type": "number"
          },
          "prixUnitaire": {
            "type": "number"
          },
          "idEntreprise": {
            "type": "integer",
            "format": "int32"
          },
          "article": {
            "$ref": "#/components/schemas/ArticleDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "JWT token",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}